+++ changelog.md
@@ -4,7 +4,9 @@ All notable changes to Formly are documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 1.5.2 - 2026-07-27
+
+[Compare changes](https://github.com/princeofscale/Formly/compare/v1.5.1...v1.5.2)
 
 ### Added